Refactored stored types to remove some narrowing warnings on x64 builds

+++ src/MemorySystem/Allocator/Growing/GrowingPoolAllocator.h
@@ -27,10 +27,10 @@ namespace sp
 			char* m_physicalMemoryEnd;
 			char* m_firstChunkPtr;
 
-			const uint32_t m_maxElementSize;
-			const uint32_t m_maxElementAlignment;
-			const uint32_t m_minimalChunkSize;
-			const uint32_t m_growSize;
+			const size_t m_maxElementSize;
+			const size_t m_maxElementAlignment;
+			const size_t m_minimalChunkSize;
+			const size_t m_growSize;
 
 			core::FreeList m_freeList;
 		};
