+++ src/MemorySystem/Allocator/NonGrowing/StackAllocator.cpp
@@ -48,10 +48,7 @@ void* sp::memory::StackAllocator::Alloc(size_t size, size_t alignment, size_t of
 {
 	assert(pointerUtil::IsPowerOfTwo(alignment) && "Alignment has to be a power-of-two");
 
-	// INFO: Used a ptrdiff_t here earlier by this would take up 8 byte on 64bit machines
-	// so I decided onto uint32_t and assume the allocator will never use more than 2^32
-	// bytes (around 4GB allowed per allocator)
-	const uint32_t allocationOffset = m_currentPtr - m_memoryBegin;
+	const ptrdiff_t allocationOffset = m_currentPtr - m_memoryBegin;
 	// Align the currentPtr offsetted by offset + AllocationMetaSize to not
 	// mess up alignment when adding canaries later
 	m_currentPtr += offset + ALLOCATION_META_SIZE;
@@ -72,7 +69,7 @@ void* sp::memory::StackAllocator::Alloc(size_t size, size_t alignment, size_t of
 
 	// Write the allocationOffset in the slot before the userPointer
 	as_char = m_currentPtr;
-	*as_uint32_t = allocationOffset;
+	*as_uint32_t = static_cast<uint32_t>(allocationOffset);
 #ifdef STACK_ALLOC_LIFO_CHECKS
 	as_char += sizeof(uint32_t);
 	*as_uint32_t = ++m_allocationID;
